Updates vig_jax.ipynb Colab, added Mixer.
Unfortunately there is still an ongoing regression in the Colab TPU Node
setup. The Mixer-B/16 works with both GPU and TPU runtimes, but the
ViT-B/32 currently only works with GPU runtime and hangs with the TPU
runtime (note that previously both ViT-B/32 and ViT-B/16 worked with
both runtimes).

Added a `NOTE` to the top of the Colab to highlight this, while waiting
for the underlying error to be fixed...